requirements.txt

Flask==2.0.1
Werkzeug==2.0.3
requests==2.28.1
dnspython==2.6.1
fpdf==1.7.2
